put things in views
