+++ src/components/cart.js
@@ -4,7 +4,7 @@ import CartToggle from './toggle';
 import Template from '../template';
 import Checkout from './checkout';
 import formatMoney from '../utils/money';
-import CartContainer from '../containers/cart';
+import CartView from '../views/cart';
 import {addClassToElement} from '../utils/element-class';
 import CartUpdater from '../updaters/cart';
 
@@ -26,7 +26,6 @@ export default class Cart extends Component {
     this.addVariantToCart = this.addVariantToCart.bind(this);
     this.childTemplate = new Template(this.config.lineItem.templates, this.config.lineItem.contents, this.config.lineItem.order);
     this.node = config.node || document.body.appendChild(document.createElement('div'));
-    this.node.className = 'shopify-buy-cart-wrapper';
     this.isVisible = this.options.startOpen;
     this.checkout = new Checkout(this.config);
     const toggles = this.globalConfig.toggles || [{
@@ -36,7 +35,7 @@ export default class Cart extends Component {
       return new CartToggle(merge({}, config, toggle), Object.assign({}, this.props, {cart: this}));
     });
     this.updater = new CartUpdater(this);
-    this.container = new CartContainer(this);
+    this.view = new CartView(this);
   }
 
   createToggles(config) {
@@ -134,10 +133,6 @@ export default class Cart extends Component {
     return this.props.client.fetchRecentCart();
   }
 
-  wrapTemplate(html) {
-    return `<div class="${this.classes.cart.cart}">${html}</div>`;
-  }
-
   /**
    * initializes component by creating model and rendering view.
    * Creates and initalizes toggle component.
@@ -153,15 +148,6 @@ export default class Cart extends Component {
       }).then(() => this);
   }
 
-  /**
-   * renders string template using viewData to wrapper element.
-   * Sets iframe class based on visibility.
-   */
-  render() {
-    super.render();
-    this.container.render();
-  }
-
   destroy() {
     super.destroy();
     this.toggles.forEach((toggle) => toggle.destroy());
@@ -172,7 +158,7 @@ export default class Cart extends Component {
    */
   close() {
     this.isVisible = false;
-    this.render();
+    this.view.render();
   }
 
   /**
@@ -180,8 +166,8 @@ export default class Cart extends Component {
    */
   open() {
     this.isVisible = true;
-    this.render();
-    this.setFocus();
+    this.view.render();
+    this.view.setFocus();
   }
 
   /**
@@ -190,9 +176,9 @@ export default class Cart extends Component {
    */
   toggleVisibility(visible) {
     this.isVisible = visible || !this.isVisible;
-    this.render();
+    this.view.render();
     if (this.isVisible) {
-      this.setFocus();
+      this.view.setFocus();
     }
   }
 
@@ -229,11 +215,11 @@ export default class Cart extends Component {
     this._userEvent('updateItemQuantity');
     return this.model.updateLineItem(id, qty).then((cart) => {
       this.model = cart;
-      this.toggles.forEach((toggle) => toggle.render());
+      this.toggles.forEach((toggle) => toggle.view.render());
       if (qty > 0) {
-        this.render();
+        this.view.render();
       } else {
-        this._animateRemoveItem(id);
+        this.view.animateRemoveNode(id);
       }
       return cart;
     });
@@ -250,9 +236,9 @@ export default class Cart extends Component {
     }
     this.open();
     return this.model.createLineItemsFromVariants({variant, quantity}).then((cart) => {
-      this.render();
-      this.toggles.forEach((toggle) => toggle.render());
-      this.setFocus();
+      this.view.render();
+      this.toggles.forEach((toggle) => toggle.view.render());
+      this.view.setFocus();
       return cart;
     });
   }
@@ -262,8 +248,8 @@ export default class Cart extends Component {
    */
   empty() {
     return this.model.clearLineItems().then(() => {
-      this.render();
-      this.toggles.forEach((toggle) => toggle.render());
+      this.view.render();
+      this.toggles.forEach((toggle) => toggle.view.render());
       return;
     });
   }
@@ -281,25 +267,5 @@ export default class Cart extends Component {
       prevQuantity: item.quantity,
       quantity: parseFloat(quantity),
     };
-  }
-
-  _animateRemoveItem(id) {
-    const el = this.container.document.getElementById(id);
-    addClassToElement('is-hidden', el);
-    if (this.props.browserFeatures.animation) {
-      el.addEventListener('animationend', () => {
-        if (!el.parentNode) {
-          return;
-        }
-        this._removeItem(el);
-      });
-    } else {
-      this._removeItem(el);
-    }
-  }
-
-  _removeItem(el) {
-    el.parentNode.removeChild(el);
-    this.render();
   }
 }
