+++ test/unit/component.js
@@ -90,7 +90,7 @@ describe('Component class', () => {
 
   describe('init', () => {
     it('fetches and renders data', () => {
-      const setupView = sinon.stub(component.frame, 'init').returns(Promise.resolve());
+      const setupView = sinon.stub(component.view, 'init').returns(Promise.resolve());
       const setupModel = sinon.stub(component, 'setupModel').returns(Promise.resolve({ title: 'test' }));
       const render = sinon.stub(component, 'render');
       const delegateEvents = sinon.stub(component, 'delegateEvents');
@@ -131,7 +131,7 @@ describe('Component class', () => {
       return testComponent.init({}).then(() => {
         testComponent.render();
         testComponent.delegateEvents();
-        testComponent.frame.document.getElementById('button').click();
+        testComponent.view.document.getElementById('button').click();
         assert.calledWith(clickSpy, sinon.match.instanceOf(Event), sinon.match.instanceOf(window.Node));
       });
     });
@@ -151,7 +151,7 @@ describe('Component class', () => {
           'product');
         const setupModel = sinon.stub(iframeComponent, 'setupModel').returns(Promise.resolve({ title: 'test' }));
         return iframeComponent.init().then(() => {
-          assert.isOk(iframeComponent.frame.iframe);
+          assert.isOk(iframeComponent.view.iframe);
           setupModel.restore();
         });
       });
